+++ TechTreeCreator/GraphCreation/BuildingGraphCreator.cs
@@ -20,9 +20,9 @@ namespace TechTreeCreator.GraphCreation {
             result.Category = node.GetKeyValue("category");
             
             node.ActOnNodes("resources", cwNode => {
-                cwNode.ActOnNodes("cost", costNode => costNode.KeyValues.ForEach(value => result.Cost[value.Key] = value.Value.ToInt()));
-                cwNode.ActOnNodes("upkeep", costNode => costNode.KeyValues.ForEach(value => result.Upkeep[value.Key] = value.Value.ToInt()));
-                cwNode.ActOnNodes("produces", costNode => costNode.KeyValues.ForEach(value => result.Produces[value.Key] = value.Value.ToInt()));
+                cwNode.ActOnNodes("cost", costNode => costNode.KeyValues.ForEach(value => result.Cost[value.Key] = value.Value.ToDouble()));
+                cwNode.ActOnNodes("upkeep", costNode => costNode.KeyValues.ForEach(value => result.Upkeep[value.Key] = value.Value.ToDouble()));
+                cwNode.ActOnNodes("produces", costNode => costNode.KeyValues.ForEach(value => result.Produces[value.Key] = value.Value.ToDouble()));
             });
         }
